+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ---
 
+## [1.6.0]
+### Added
+- 
+
+### Changed
+- Neo4j data directories moved => data/data, data/logs and data/plugins moved into data/database directory
+- No user used to launch docker compose to improve compatibility with external SSD
+
+### Fixed
+- 
+
 
 ## [1.5.0]
 ### Added